Basic Camera Controls
Very basic player and camera movement for the game. For now, the player
moves using WASD as the camera follows the character. More advanced
camera controls will be added later in the project. Also committing a
text file with instructions on how to set up the camera and the
controls.